move XmlUtilities to util folder
