+++ src/main/java/org/opengis/cite/cdb10/MetadataAndVersioning/VendorAttributesXmlStructureTests.java
@@ -1,8 +1,8 @@
 package org.opengis.cite.cdb10.MetadataAndVersioning;
 
-import org.opengis.cite.cdb10.CDBStructure.XmlUtilities;
 import org.opengis.cite.cdb10.CommonFixture;
 import org.opengis.cite.cdb10.util.SchemaValidatorErrorHandler;
+import org.opengis.cite.cdb10.util.XmlUtilities;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.xml.sax.SAXException;
